% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{agrupaciones}
\alias{agrupaciones}
\alias{agrupaciones}
\title{Agrupaciones que participaron en las Paso 2019}
\format{Un data frame con 182 filas y 2 variables:
\describe{
  \item{CODIGO_AGRUPACION}{Código de la agrupación política}
  \item{NOMBRE_AGRUPACION}{Nombre de la agrupación política}
}}
\source{
\url{http://descargaresultados.s3-sa-east-1.amazonaws.com/resultados.zip}

\url{http://descargaresultados.s3-sa-east-1.amazonaws.com/resultados.zip}
}
\usage{
agrupaciones

agrupaciones
}
\description{
Un dataset conteniendo el detalle de todas las agrupaciones de esta elección,
deriva de descripcion_postulaciones.dsv

Un dataset conteniendo el detalle de todas las mesas de votación de esta elección,
deriva de descripcion_postulaciones.dsv
}
\keyword{datasets}
