% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{categorias}
\alias{categorias}
\title{Categorias a las que se votó}
\format{Un data frame con 204 filas y 4 variables:
\describe{
  \item{id_categoria}{id numérico generado}
  \item{codigo_categoria}{Código del categoría del voto}
  \item{nombre_categoria}{Nombre de la categoría}
  \item{votos_totales}{Sumatoria de todos los votos en esta categoría}
}}
\usage{
categorias
}
\description{
Un dataset conteniendo el detalle de todas las categorias de esta elección,
deriva de descripcion_postulaciones.dsv. La categoría representa lo que
se está votando.
}
\details{
(MODELO NUEVO)
}
\keyword{datasets}
