% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{circuitos}
\alias{circuitos}
\title{Circuitos electorales}
\format{Un data frame con 5,649 filas y 3 variables:
\describe{
  \item{id_circuito}{un id numérico generado para el circuito}
  \item{codigo_circuito}{código del circuito de la mesa}
  \item{nombre_circuito}{nombre del circuito}
}}
\usage{
circuitos
}
\description{
Un dataset conteniendo el detalle de los circuitos electorales,
deriva de mesas_totales_lista.dsv
}
\keyword{datasets}
