% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{distritos}
\alias{distritos}
\title{Distritos electorales (Provincias)}
\format{Un data frame con 24 filas y 3 variables:
\describe{
  \item{id_distrito}{un id numérico generado para el distrito}
  \item{codigo_distrito}{código de la provincia}
  \item{nombre_distrito}{Nombre de la provincia}
}}
\usage{
distritos
}
\description{
Un dataset conteniendo el detalle de los distritos electorales,
deriva de mesas_totales_lista.dsv
}
\keyword{datasets}
