% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{establecimientos}
\alias{establecimientos}
\title{establecimientos}
\format{Un data frame con 15,226 filas y 6 variables:
\describe{
  \item{id_establecimiento}{id del establecimiento}
  \item{codigo_establecimiento}{El c贸digo del establecimiento}
  \item{nombre_establecimiento}{Nombre del establecimiento}
  \item{codigo_mesa}{c贸digo de la mesa tal como parece en los datos originales}
  \item{id_circuito}{id del circuito}
  \item{id_seccion}{id de la secci贸n}
  \item{id_distrito}{id del distrito}
}}
\usage{
establecimientos
}
\description{
Datos de establecimientos o locales de votaci贸n procesados a partir del
dataset scrap_establecimientos_mesas.
}
\keyword{datasets}
