% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_telegrama_url.R
\name{get_telegrama_url}
\alias{get_telegrama_url}
\title{Get the url path to the "telegram" of a particular "mesa"}
\usage{
get_telegrama_url(mesa,
  base_url = "https://www.resultados2019.gob.ar/opt/jboss/rct/tally/pages")
}
\arguments{
\item{mesa}{The "mesa" code.}

\item{base_url}{A base path, default: "https://www.resultados2019.gob.ar/opt/jboss/rct/tally/pages".}
}
\value{
A url.
}
\description{
Get the url path to the "telegram" of a particular "mesa"
}
\examples{
get_telegrama_url("0100100001X")
get_telegrama_url("0100100001X", base_url = "https://www.resultados2019.gob.ar/opt/jboss/rct/tally/pages")

}
