% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{listas}
\alias{listas}
\title{Listas a las que se votó}
\format{Un data frame con 1442 filas y 6 variables:
\describe{
  \item{id_lista}{id numérico generado}
  \item{id_agrupación}{id a la agrupación de la lista}
  \item{código_lista}{código tal como aparece en los archivos originales}
  \item{nombre_lista}{nombre de la lista}
}}
\usage{
listas
}
\description{
Un dataset conteniendo el detalle de todas las listas de esta elección,
deriva de descripcion_postulaciones.dsv. Un Partido puede tener múltiples
listas, las PASO justamente buscan definir estas.
}
\details{
(MODELO NUEVO)
}
\keyword{datasets}
