% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{mesas}
\alias{mesas}
\title{Mesas totales}
\format{Un data frame con 100,148 filas y 2 variables:
\describe{
  \item{id_mesa}{id numérico generado}
  \item{id_distrito}{id del distrito de la mesa (provincia)}
  \item{id_seccion}{id del la sección electoral}
  \item{id_circuito}{id del circuito}
  \item{codigo_mesa}{código de la mesa tal como parece en los datos originales}
  \item{escrutada}{la mesa fue escrutada}
}}
\usage{
mesas
}
\description{
Un dataset conteniendo el detalle de todas las mesas de votación de esta elección,
deriva de mesas_totales_lista.dsv
}
\keyword{datasets}
