% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{mesas_totales_agrp_politica}
\alias{mesas_totales_agrp_politica}
\title{Votos por agrupación}
\format{Un data frame con 3,665,201 filas y 8 variables:
\describe{
  \item{CODIGO_DISTRITO}{Código de distrito}
  \item{CODIGO_SECCION}{Código de sección electoral}
  \item{CODIGO_CIRCUITO}{Código de circuito}
  \item{CODIGO_MESA}{Código de mesa}
  \item{CODIGO_CATEGORIA}{Código de categoría de la elección}
  \item{CODIGO_AGRUPACION}{Código de agrupación}
}}
\source{
\url{http://descargaresultados.s3-sa-east-1.amazonaws.com/resultados.zip}
}
\usage{
mesas_totales_agrp_politica
}
\description{
Un dataset conteniendo el detalle de los totales de votos por mesa, abierto hasta la agrupación,
derivado del archivo original mesas_totales_agrp_politica.dsv
}
\details{
(ESTRUCTURA ORIGINAL)
}
\keyword{datasets}
