% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{meta_agrupaciones}
\alias{meta_agrupaciones}
\title{meta_agrupaciones}
\format{Un data frame con 101 filas y 2 variables:
\describe{
  \item{id_meta_agrupacion}{id numérico generado}
  \item{nombre_meta_agrupacion}{Nombre del partido}
}}
\usage{
meta_agrupaciones
}
\description{
Un dataset conteniendo el detalle de todos los partidos. En las tablas originales
existen agrupaciones, las cuales se corresponden con los partidos, pero tienen
distinto código dependiendo de la categoria  (Presidente, Gobernador, Senadores, etc)
y del distrito. Por lo que agregamos el concepto de meta_agrupacion que
efectivamete representa al partido.
}
\details{
deriva de: descripcion_postulaciones.dsv

(MODELO NUEVO)
}
\keyword{datasets}
