% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dsv_and_create_model.R
\name{process_dsv_and_create_model}
\alias{process_dsv_and_create_model}
\title{Procesa los dsv de las paso 2019 y genera el modelo de datos.
No es necesario correrlo ya que el modelo completo ya se distribuye en
el paquete.}
\usage{
process_dsv_and_create_model()
}
\description{
Procesa los dsv de las paso 2019 y genera el modelo de datos.
No es necesario correrlo ya que el modelo completo ya se distribuye en
el paquete.
}
