% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{scrap_establecimientos_mesas}
\alias{scrap_establecimientos_mesas}
\title{scrap_establecimientos_mesas}
\format{Un data frame con 101,457 filas y 4 variables:
\describe{
  \item{codigo_establecimiento}{El código del establecimiento}
  \item{codigo_circuito}{código del circuito del establecimiento}
  \item{nombre_establecimiento}{Nombre del establecimiento}
  \item{codigo_mesa}{código de la mesa tal como parece en los datos originales}
}}
\usage{
scrap_establecimientos_mesas
}
\description{
Datos de establecimientos o locales de votación  y mesas. Esta información
se capturó leyendo directamente los archivos JSON de datos que maneja la
aplicación web.
}
\keyword{datasets}
