% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{secciones}
\alias{secciones}
\title{Secciones electorales (municipios)}
\format{Un data frame con 525 filas y 3 variables:
\describe{
  \item{id_seccion}{un id numérico generado para la sección}
  \item{codigo_seccion}{código de la sección}
  \item{nombre_seccion}{Nombre de la sección}
}}
\usage{
secciones
}
\description{
Un dataset conteniendo el detalle de las secciones electorales,
deriva de mesas_totales_lista.dsv
}
\keyword{datasets}
