% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_telegrama.R
\name{view_telegrama}
\alias{view_telegrama}
\title{Ver el telegrama de una determinada mesa}
\usage{
view_telegrama(codigo_mesa,
  base_url = "https://www.resultados2019.gob.ar/opt/jboss/rct/tally/pages")
}
\arguments{
\item{codigo_mesa}{El código de mesa}

\item{base_url}{El path base, default: "https://www.resultados2019.gob.ar/opt/jboss/rct/tally/pages".}
}
\value{
una url.
}
\description{
Ver el telegrama de una determinada mesa
}
\examples{

library("paso2019")

view_telegrama("0100100002X")

}
