% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{votos}
\alias{votos}
\title{Votos}
\format{Un data frame con 4,021,306 filas y 5 variables:
\describe{
  \item{id_voto}{un id numérico generado para el conteo de votos}
  \item{id_mesa}{id de la mesa}
  \item{id_categoria}{id de la categoría de los votos}
  \item{id_lista}{id de la lista}
  \item{votos}{Votos totales}
}}
\usage{
votos
}
\description{
Un dataset conteniendo el detalle de los votos totales,
deriva de mesas_totales_lista.dsv
}
\keyword{datasets}
