% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{agrupaciones}
\alias{agrupaciones}
\title{Agrupaciones que participaron}
\format{Un data frame con 182 filas y 2 variables:
\describe{
  \item{id_agrupacion}{id numérico generado}
  \item{id_meta_agrupacion}{id a la meta agrupación o partido}
  \item{codigo_agrupacion}{código de la agrupación tal como se informa}
}}
\usage{
agrupaciones
}
\description{
Un dataset conteniendo el detalle de todas las agrupaciones de esta elección,
deriva de descripcion_postulaciones.dsv
}
\keyword{datasets}
