% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{categorias}
\alias{categorias}
\title{categorias a las que se votó en las Paso 2019
(MODELO NUEVO)
Un dataset conteniendo el detalle de todas las categorias de esta elección,
deriva de descripcion_postulaciones.dsv}
\format{Un data frame con 1442 filas y 6 variables:
\describe{
  \item{id_categoria}{id numérico generado}
  \item{codigo_categoria}{Código del categoría del voto}
  \item{nombre_categoria}{Nombre de la categoría}
  \item{votos_totales}{Sumatoria de todos los votos en esta categoría}
}}
\source{
\url{http://descargaresultados.s3-sa-east-1.amazonaws.com/resultados.zip}
}
\usage{
categorias
}
\description{
categorias a las que se votó en las Paso 2019
(MODELO NUEVO)
Un dataset conteniendo el detalle de todas las categorias de esta elección,
deriva de descripcion_postulaciones.dsv
}
\keyword{datasets}
