% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{descripcion_regiones}
\alias{descripcion_regiones}
\title{Regiones electorales Argentinas}
\format{Un data frame con 6,298 filas y 2 variables:
\describe{
  \item{CODIGO_REGION}{Código de la región}
  \item{NOMBRE_REGION}{Nombre de la región}
}}
\source{
\url{http://descargaresultados.s3-sa-east-1.amazonaws.com/resultados.zip}
}
\usage{
descripcion_regiones
}
\description{
Un dataset conteniendo los nombres de las regiones electorales,
tal como están definidas  en el archivo original descripcion_regiones.dsv.
En un mismo archivo están los 3 niveles: distrito, seccion y
circuito el CODIGO_REGION es particular para cada una de
estas entidades.
}
\details{
(ESTRUCTURA ORIGINAL)
}
\keyword{datasets}
