% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_telegrama_url.R
\name{get_telegrama_url}
\alias{get_telegrama_url}
\title{Obtener la url de la imagen digitalizada del telegrama}
\usage{
get_telegrama_url(codigo_mesa,
  base_url = "https://www.resultados2019.gob.ar/opt/jboss/rct/tally/pages")
}
\arguments{
\item{codigo_mesa}{El código de mesa}

\item{base_url}{El path base, default: "https://www.resultados2019.gob.ar/opt/jboss/rct/tally/pages".}
}
\value{
A url.
}
\description{
Obtener la url de la imagen digitalizada del telegrama
}
\examples{

get_telegrama_url("0100100001X")
get_telegrama_url("0100100001X", base_url = "https://www.resultados2019.gob.ar/opt/jboss/rct/tally/pages")

}
