% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paso2019.R
\docType{data}
\name{mesas_totales_lista}
\alias{mesas_totales_lista}
\title{Votos por mesa y lista}
\format{Un data frame con 3665201 filas y 8 variables:
\describe{
  \item{CODIGO_DISTRITO}{Código de distrito}↕
  \item{CODIGO_SECCION}{Código de sección electoral}
  \item{CODIGO_CIRCUITO}{Código de circuito}
  \item{CODIGO_MESA}{Código de mesa}
  \item{CODIGO_CATEGORIA}{Código de categoría de la elección}
  \item{CONTADOR}{Sin información disponible}
  \item{VALOR}{sin información disponible}
}}
\source{
\url{http://descargaresultados.s3-sa-east-1.amazonaws.com/resultados.zip}
}
\usage{
mesas_totales_lista
}
\description{
Un dataset conteniendo el detalle de todas las agrupaciones que se han presentado,
derivado del archivo original mesas_totales_lista.dsv
}
\details{
(ESTRUCTURA ORIGINAL)
}
\keyword{datasets}
